+++ web/hooks/index.ts
@@ -1,7 +1,6 @@
 export * from './spl-token';
 export * from './store';
 export * from './games';
-export * from './use-count-down';
 
 export * from '@/components/ui/ui-layout';
 export * from '@/components/account/account-data-access';