+++ web/providers/index.ts
@@ -1,12 +1,14 @@
 export * from './bg-provider';
 export * from './game-pal';
 export * from './messages';
+export * from './countdown';
 
 export * from './react-query-provider';
 export * from './crypto-provider';
 export * from './lucky-bags-provider';
 export * from './wallet-provider';
 export * from './solana-provider';
+export * from './bag-provider';
 
 export * from '@/components/cluster/cluster-data-access';
 export * from '@/app/console/games/_provider/games-provider';