+++ web/ui/index.ts
@@ -1,2 +1,4 @@
 export * from './Logo';
 export * from './messages';
+export * from './CountdownBag';
+export * from './BagKeyForm';