+++ CHANGELOG.md
@@ -2,6 +2,43 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.11](///compare/v0.1.10...v0.1.11) (2024-07-16)
+
+
+### Features
+
+* **i18n:** 👻 Bag | Messages 39e24c7
+* **i18n:** 👻 Connect | External Wallet ff14d30
+* **web:** 👻 Bag | Messages 6de0e0d
+* **web:** 👻 Generated | Messages 55a70b9
+* **web:** 👻 Locked | Messages a4338ec
+* **web:** 👻 LuckyWallet | Context ad0e22e
+* **web:** 👻 Selector | Messages e5d40a5
+
+### [0.1.10](///compare/v0.1.9...v0.1.10) (2024-07-15)
+
+
+### Features
+
+* **i18n:** 👻 Delete | LuckyBag 9b24a5b
+* **i18n:** 👻 Lobby 5357aa2
+* **i18n:** 👻 LuckyBag | Messages 638df11
+* **web:** 👻 Activate | Message c4215ea
+* **web:** 👻 ChatController | Component 04f4197
+* **web:** 👻 Countdown | Context 2ec5a60
+* **web:** 👻 CountdownBag | Component 0f2addc
+* **web:** 👻 Encryption | LuckyBags 8e2b844
+* **web:** 👻 Flow | ChatController 840946d
+* **web:** 👻 Generate | Message 77ed748
+* **web:** 👻 Lobby | group path 0c86e71
+* **web:** 👻 Lobby | Layout 2355627
+* **web:** 👻 Locked | Message 8c0be5e
+* **web:** 👻 Password Form | LuckyBags 7e6e524
+* **web:** 👻 Path | ChatController b92abd1
+* **web:** 👻 Secure | Message 354f875
+* **web:** 👻 Selector | Message a6e59ac
+* **web:** 👻 Stored Key | Crypto 6bef9c1
+
 ### [0.1.9](///compare/v0.1.8...v0.1.9) (2024-07-14)